+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- New fetcher support: AWS S3
+  - Retrieve certs from S3 buckets and S3-like storage
+
 ---
 
 ## Stable Releases
@@ -57,6 +62,49 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ---
 
 ## Pre-Releases
+
+## [0.2.2-beta.5] - 2025-11-06
+
+### Added
+
+- Nada - just testing the new release testing workflow
+
+---
+
+## [0.2.2-beta.4] - 2025-11-04
+
+### Added
+
+- New fetcher support: Google Cloud Storage!
+  - Retrieve certs from gcs accounts
+
+---
+
+## [0.2.2-beta.3] - 2025-11-04
+
+### Added
+
+- New fetcher support: Azure Blob!
+  - Retrieve certs from azure blob storage accounts
+
+---
+
+## [0.2.2-beta.2] - 2025-11-05
+
+### Fixed
+
+- Added s3 fetch type to schema validation
+
+---
+
+## [0.2.2-beta.1] - 2025-11-04
+
+### Added
+
+- New fetcher support: AWS S3
+  - Retrieve certs from S3 buckets and S3-like storage
+
+---
 
 ## [0.1.3-beta.32] - 2025-11-03